+++ CHANGELOG.md
@@ -77,6 +77,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * `jj gerrit upload` now correctly handles mixed explicit and implicit
   Change-Ids in chains of commits ([#8219](https://github.com/jj-vcs/jj/pull/8219))
 
+* Working copies with symlink loops clone correctly
+
 ## [0.36.0] - 2025-12-03
 
 ### Release highlights